# dapp/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ritoswap OpenAPI",
    "version": "0.3.0",
    "description": "RitoSwap API - Token-gated access and NFT verification endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current environment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
    "/api/token-status/{tokenId}": {
      "get": {
        "summary": "Get token status and usage information",
        "description": "Checks if a token exists on-chain and whether it has been used for gated access. This endpoint is optimized for polling with light caching.",
        "tags": [
          "Token Status",
          "NFT"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "description": "The NFT token ID to check",
              "example": "123"
            },
            "required": true,
            "description": "The NFT token ID to check",
            "name": "tokenId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Token status retrieved successfully",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Cache control header for successful responses"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "exists",
                    "used",
                    "usedBy",
                    "usedAt"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1,
                      "description": "Number of tokens found (0 or 1)"
                    },
                    "exists": {
                      "type": "boolean",
                      "description": "Whether the token exists on-chain"
                    },
                    "used": {
                      "type": "boolean",
                      "description": "Whether the token has been used for gated access"
                    },
                    "usedBy": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Address that used the token (if used)"
                    },
                    "usedAt": {
                      "oneOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "ISO timestamp when token was used (if used)"
                    }
                  }
                },
                "examples": {
                  "exists-used": {
                    "summary": "Token exists and has been used",
                    "value": {
                      "count": 1,
                      "exists": true,
                      "used": true,
                      "usedBy": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb7",
                      "usedAt": "2025-01-15T10:30:00.000Z"
                    }
                  },
                  "exists-unused": {
                    "summary": "Token exists but not used",
                    "value": {
                      "count": 1,
                      "exists": true,
                      "used": false,
                      "usedBy": null,
                      "usedAt": null
                    }
                  },
                  "not-exists": {
                    "summary": "Token does not exist",
                    "value": {
                      "count": 0,
                      "exists": false,
                      "used": false,
                      "usedBy": null,
                      "usedAt": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token ID provided",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "enum": [
                        "https://ritoswap.com/errors/invalid-token-id"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "enum": [
                        "Invalid token ID"
                      ]
                    },
                    "status": {
                      "type": "number",
                      "enum": [
                        400
                      ]
                    },
                    "detail": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "description": "Invalid token ID provided",
                  "example": {
                    "type": "https://ritoswap.com/errors/invalid-token-id",
                    "title": "Invalid token ID",
                    "status": 400,
                    "detail": "Token ID must be a non-negative integer"
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "description": "Method not allowed",
                  "example": {
                    "error": "Method not allowed"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "string",
                  "example": "60"
                },
                "required": true
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "string",
                  "example": "0"
                },
                "required": true
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "example": "60"
                },
                "required": true
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "no-store, no-cache, must-revalidate, max-age=0, s-maxage=0"
                },
                "required": true
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "remaining": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "retryAfter": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "additionalProperties": true,
                  "description": "Rate limit exceeded",
                  "example": {
                    "type": "https://ritoswap.com/errors/too-many-requests",
                    "title": "Too many requests",
                    "status": 429,
                    "detail": "Rate limit exceeded for token-status",
                    "limit": 60,
                    "remaining": 0,
                    "retryAfter": 60
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "remaining": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "retryAfter": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "additionalProperties": true,
                  "description": "Internal server error",
                  "example": {
                    "type": "https://ritoswap.com/errors/failed-to-check-token-status",
                    "title": "Failed to check token status",
                    "status": 500,
                    "detail": "An unexpected error occurred while checking token status"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}